Renamed components with underscores in their names
